pyproject.toml
[tool.poetry]
name = "dr-hatespeech"
version = "0.0.1"
description = "Hatespeech detection for DR Facebook data."
authors = [
    "Dan Saattrup Nielsen <dan.nielsen@alexandra.dk>",
    "Anders Jess Pedersen <anders.j.pedersen@alexandra.dk>",
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
hydra-core = "^1.2.0"
datasets = "^2.3.2"
fastparquet = "^0.8.1"
protobuf = "^4.21.2"
transformers = "^4.20.1"
torch = "^1.12.0"
label-studio = "^1.5.0"
jupyterlab = "^3.4.3"
ipywidgets = "^7.7.1"
matplotlib = "^3.5.2"
snorkel = {git = "https://github.com/saattrupdan/snorkel", rev = "main"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pdoc = "^7.1.1"
pytest = "^7.1.2"
pre-commit = "^2.17.0"
pytest-xdist = "^2.5.0"
pytest-cov = "^3.0.0"
readme-coverage-badger = "^0.1.2"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    '--verbose',
    '--durations=10',
    '--color=yes',
    '-s',
    '-vv',
    '--doctest-modules',
    '-n 2',
    '--cov=src/dr_hatespeech',
]
xfail_strict = true
filterwarnings = ["ignore"]
log_cli_level = "info"
testpaths = ["tests", "src/dr_hatespeech"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.isort]
profile = "black"
